+++ PyGeopack/MLONtoMLT.py
@@ -18,11 +18,11 @@ def MLONtoMLT(MLon, Date, ut):
 	
 	'''
 	#Convert input variables to appropriate numpy dtype:
-	_MLon = np.array([MLon]).flatten().astype("float32")
+	_MLon = np.array([MLon]).flatten().astype("float64")
 	_n = np.int32(_MLon.size)
 	_date = np.int32(Date)
 	_UT = np.float32(ut)
-	_MLT = np.zeros(_n,dtype="float32")
+	_MLT = np.zeros(_n,dtype="float64")
 	_CMLONtoMLTUT(_MLon, _n, _date, _UT, _MLT)
 
 	return _MLT
